pyproject.toml
[project]
name = "goiania-housing-market-prediction"
version = "0.1.0"
description = ""
authors = [
    {name = "AndreKoraleski",email = "andrekorale@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
]

[build-system]
requires = ["hatchling>=1.13,<2"]
build-backend = "hatchling.build"

[tool.hatch.build]

[tool.hatch.build.targets.wheel]
packages = ["goh"]
